requirements.txt
requests
datetime
pandas
numpy
gunicorn
Flask
google-cloud-storage==2.1.0
